+++ src/twigextensions/ReadTimeTwigExtension.php
@@ -1,6 +1,6 @@
 <?php
 /**
- * Read Time plugin for Craft CMS 3.x
+ * Read Time plugin for Craft CMS 4.x
  *
  * Calculate the estimated read time for content.
  *
@@ -10,53 +10,60 @@
 
 namespace jalendport\readtime\twigextensions;
 
+use benf\neo\elements\Block as NeoBlock;
+use benf\neo\Field as NeoField;
+use Craft;
+use craft\elements\Entry;
+use craft\elements\MatrixBlock;
+use craft\errors\InvalidFieldException;
+use craft\fields\Matrix;
+use craft\helpers\StringHelper;
+use jalendport\readtime\models\Settings;
 use jalendport\readtime\ReadTime;
 use jalendport\readtime\models\TimeModel;
-
-use Craft;
-use craft\helpers\DateTimeHelper;
-use craft\helpers\StringHelper;
-
+use Twig\Extension\AbstractExtension;
+use Twig\TwigFilter;
+use Twig\TwigFunction;
+use verbb\supertable\fields\SuperTableField;
 use yii\base\ErrorException;
 
-class ReadTimeTwigExtension extends \Twig_Extension
+class ReadTimeTwigExtension extends AbstractExtension
 {
     // Public Methods
     // =========================================================================
 
-    public function getName()
-    {
+    public function getName(): string
+	{
         return 'readTime';
     }
 
-    public function getFunctions()
-    {
+    public function getFunctions(): array
+	{
         return [
-            new \Twig_SimpleFunction('readTime', [$this, 'readTimeFunction']),
+            new TwigFunction('readTime', [$this, 'readTimeFunction']),
         ];
     }
 
-    public function getFilters()
-    {
+    public function getFilters(): array
+	{
         return [
-            new \Twig_SimpleFilter('readTime', [$this, 'readTimeFilter']),
+            new TwigFilter('readTime', [$this, 'readTimeFilter']),
         ];
     }
 
-    public function readTimeFunction($element, $showSeconds = true)
-    {
+	public function readTimeFunction($element, $showSeconds = true): TimeModel
+	{
         $totalSeconds = 0;
-        $vals = '';
 
-        if ($element instanceof \craft\elements\Entry) {
+        if ($element instanceof Entry) {
             // Provided value is an entry
 
-            foreach ($element->getFieldLayout()->getFields() as $field) {
+            foreach ($element->getFieldLayout()->getCustomFields() as $field) {
                 try {
-                    // If field is a matrix then loop through fields in block
-                    if ($field instanceof \craft\fields\Matrix || $field instanceof \benf\neo\Field) {
+                    // If field is a matrix or neo field then loop through fields in block
+                    if ($field instanceof Matrix || $field instanceof NeoField) {
                         foreach($element->getFieldValue($field->handle)->all() as $block) {
-                            $blockFields = $block->getFieldLayout()->getFields();
+                            $blockFields = $block->getFieldLayout()->getCustomFields();
 
                             foreach ($blockFields as $blockField) {
                                 $value = $block->getFieldValue($blockField->handle);
@@ -64,14 +71,14 @@ class ReadTimeTwigExtension extends \Twig_Extension
                                 $totalSeconds = $totalSeconds + $seconds;
                             }
                         }
-                    } elseif($field instanceof \verbb\supertable\fields\SuperTableField) {
+                    } elseif($field instanceof SuperTableField) {
                         foreach($element->getFieldValue($field->handle)->all() as $block) {
-                            $blockFields = $block->getFieldLayout()->getFields();
+                            $blockFields = $block->getFieldLayout()->getCustomFields();
 
                             foreach ($blockFields as $blockField) {
-                                if ($blockField instanceof \craft\fields\Matrix) {
+                                if ($blockField instanceof Matrix) {
                                     foreach($block->getFieldValue($blockField->handle)->all() as $matrix) {
-                                        $matrixFields = $matrix->getFieldLayout()->getFields();
+                                        $matrixFields = $matrix->getFieldLayout()->getCustomFields();
 
                                         foreach ($matrixFields as $matrixField) {
                                             $value = $matrix->getFieldValue($matrixField->handle);
@@ -91,7 +98,7 @@ class ReadTimeTwigExtension extends \Twig_Extension
                         $seconds = $this->valToSeconds($value);
                         $totalSeconds = $totalSeconds + $seconds;
                     }
-                } catch (ErrorException $e) {
+                } catch (ErrorException | InvalidFieldException $e) {
                     continue;
                 }
             }
@@ -100,8 +107,8 @@ class ReadTimeTwigExtension extends \Twig_Extension
             Craft::info('matrix or neo field provided', 'readtime');
 
             foreach ($element as $block) {
-                if ($block instanceof \craft\elements\MatrixBlock || $block instanceof \benf\neo\elements\Block) {
-                    $blockFields = $block->getFieldLayout()->getFields();
+                if ($block instanceof MatrixBlock || $block instanceof NeoBlock) {
+                    $blockFields = $block->getFieldLayout()->getCustomFields();
 
                     foreach ($blockFields as $blockField) {
                         $value = $block->getFieldValue($blockField->handle);
@@ -120,8 +127,8 @@ class ReadTimeTwigExtension extends \Twig_Extension
         return new TimeModel($data);
     }
 
-    public function readTimeFilter($value = null, $showSeconds = true)
-    {
+    public function readTimeFilter($value = null, $showSeconds = true): TimeModel
+	{
         $seconds = $this->valToSeconds($value);
 
         $data = [
@@ -135,9 +142,10 @@ class ReadTimeTwigExtension extends \Twig_Extension
     // Private Methods
     // =========================================================================
 
-    private function valToSeconds($value)
-    {
-        $settings = ReadTime::$plugin->getSettings();
+    private function valToSeconds($value): float
+	{
+		/** @var Settings $settings */
+        $settings = ReadTime::getInstance()->getSettings();
         $wpm = $settings->wordsPerMinute;
 
         $string = StringHelper::toString($value);